Moved *_utils to common/utils/*
