+++ components/src/dynamo/common/__init__.py
@@ -9,11 +9,10 @@ Dynamo backends and components.
 
 Main submodules:
     - config_dump: Configuration dumping and system diagnostics utilities
-    - env_utils: Environment variable parsing utilities
-    - prometheus_utils: Prometheus metrics collection and logging utilities
+    - utils: Common utilities including environment and prometheus helpers
 """
 
-from dynamo.common import config_dump, env_utils, prometheus_utils
+from dynamo.common import config_dump, utils
 
 try:
     from ._version import __version__
@@ -25,4 +24,4 @@ except Exception:
     except Exception:
         __version__ = "0.0.0+unknown"
 
-__all__ = ["__version__", "config_dump", "env_utils", "prometheus_utils"]
+__all__ = ["__version__", "config_dump", "utils"]
