+++ components/src/dynamo/vllm/main.py
@@ -14,7 +14,7 @@ from vllm.v1.engine.async_llm import AsyncLLM
 
 from dynamo._core import Endpoint
 from dynamo.common.config_dump import dump_config
-from dynamo.common.prometheus_utils import (
+from dynamo.common.utils.prometheus import (
     is_engine_metrics_callback_enabled,
     register_engine_metrics_callback,
 )
